requirements.txt
Flask==2.3.6
mysql-connector-python==8.1.0
Flask-Login==0.6.2
Jinja2==3.1.2
Werkzeug==2.3.6